refactor: impl Environment for ConcreteEnvironment
Use enum_dispatch to implement Environment for ConcreteEnvironment.
It makes sense for us to use enum_dispatch rather than trait objects
since:
- We have a limited number of types that implement Environment (3)
- Sometimes we want to preserve a concrete type
- There are some limitations on how trait objects can be used in
  combination with generics
